# Warm up PowerShell Worker before receiving function load request