--- a/src/RequestProcessor.cs
+++ b/src/RequestProcessor.cs
@@ -25,6 +25,7 @@ namespace Microsoft.Azure.Functions.PowerShellWorker
     internal class RequestProcessor
     {
         private readonly MessagingStream _msgStream;
+        private readonly System.Management.Automation.PowerShell _firstPwshInstance;
         private readonly PowerShellManagerPool _powershellPool;
         private DependencyManager _dependencyManager;
 
@@ -37,9 +38,10 @@ namespace Microsoft.Azure.Functions.PowerShellWorker
         private Dictionary<StreamingMessage.ContentOneofCase, Func<StreamingMessage, StreamingMessage>> _requestHandlers =
             new Dictionary<StreamingMessage.ContentOneofCase, Func<StreamingMessage, StreamingMessage>>();
 
-        internal RequestProcessor(MessagingStream msgStream)
+        internal RequestProcessor(MessagingStream msgStream, System.Management.Automation.PowerShell firstPwshInstance)
         {
             _msgStream = msgStream;
+            _firstPwshInstance = firstPwshInstance;
             _powershellPool = new PowerShellManagerPool(() => new RpcLogger(msgStream));
 
             // Host sends capabilities/init data to worker
@@ -197,15 +199,11 @@ namespace Microsoft.Azure.Functions.PowerShellWorker
                     // Setup the FunctionApp root path and module path.
                     FunctionLoader.SetupWellKnownPaths(functionLoadRequest, managedDependenciesPath);
 
-                    // Create the very first Runspace so the debugger has the target to attach to.
-                    // This PowerShell instance is shared by the first PowerShellManager instance created in the pool,
-                    // and the dependency manager (used to download dependent modules if needed).
-                    var pwsh = Utils.NewPwshInstance();
-                    LogPowerShellVersion(rpcLogger, pwsh);
-                    _powershellPool.Initialize(pwsh);
+                    LogPowerShellVersion(rpcLogger, _firstPwshInstance);
+                    _powershellPool.Initialize(_firstPwshInstance);
 
                     // Start the download asynchronously if needed.
-                    _dependencyManager.StartDependencyInstallationIfNeeded(request, pwsh, rpcLogger);
+                    _dependencyManager.StartDependencyInstallationIfNeeded(request, _firstPwshInstance, rpcLogger);
 
                     rpcLogger.Log(isUserOnlyLog: false, LogLevel.Trace, string.Format(PowerShellWorkerStrings.FirstFunctionLoadCompleted, stopwatch.ElapsedMilliseconds));
                 }