--- a/src/Utility/Utils.cs
+++ b/src/Utility/Utils.cs
@@ -37,11 +37,6 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.Utility
         {
             if (s_iss == null)
             {
-                if (FunctionLoader.FunctionAppRootPath == null)
-                {
-                    throw new InvalidOperationException(PowerShellWorkerStrings.FunctionAppRootNotResolved);
-                }
-
                 s_iss = InitialSessionState.CreateDefault();
 
                 if (!AreDurableFunctionsEnabled())
@@ -51,11 +46,14 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.Utility
                     s_iss.ThreadOptions = PSThreadOptions.UseCurrentThread;
                 }
 
-                s_iss.EnvironmentVariables.Add(
-                    new SessionStateVariableEntry(
-                        "PSModulePath",
-                        FunctionLoader.FunctionModulePath,
-                        description: null));
+                if (FunctionLoader.FunctionAppRootPath != null)
+                {
+                    s_iss.EnvironmentVariables.Add(
+                        new SessionStateVariableEntry(
+                            "PSModulePath",
+                            FunctionLoader.FunctionModulePath,
+                            description: null));
+                }
 
                 // Setting the execution policy on macOS and Linux throws an exception so only update it on Windows
                 if(Platform.IsWindows)