--- a/src/Worker.cs
+++ b/src/Worker.cs
@@ -4,10 +4,12 @@
 //
 
 using System;
+using System.Management.Automation;
 using System.Threading.Tasks;
 
 using CommandLine;
 using Microsoft.Azure.Functions.PowerShellWorker.Messaging;
+using Microsoft.Azure.Functions.PowerShellWorker.PowerShell;
 using Microsoft.Azure.Functions.PowerShellWorker.Utility;
 using Microsoft.Azure.WebJobs.Script.Grpc.Messages;
 
@@ -39,18 +41,37 @@ namespace Microsoft.Azure.Functions.PowerShellWorker
             // and the dependency manager (used to download dependent modules if needed).
             var firstPowerShellInstance = Utils.NewPwshInstance();
             LogPowerShellVersion(firstPowerShellInstance);
+            WarmUpPowerShell(firstPowerShellInstance);
 
             var msgStream = new MessagingStream(arguments.Host, arguments.Port);
             var requestProcessor = new RequestProcessor(msgStream, firstPowerShellInstance);
 
             // Send StartStream message
-            var startedMessage = new StreamingMessage() {
+            var startedMessage = new StreamingMessage()
+            {
                 RequestId = arguments.RequestId,
                 StartStream = new StartStream() { WorkerId = arguments.WorkerId }
             };
 
             msgStream.Write(startedMessage);
             await requestProcessor.ProcessRequestLoop();
+        }
+
+        // Warm up the PowerShell instance so that the subsequent function load and invocation requests are faster
+        private static void WarmUpPowerShell(System.Management.Automation.PowerShell firstPowerShellInstance)
+        {
+            // We just need this name to be unique, so that it does not coincide with an actual function
+            const string DummyFunctionName = "DummyFunction-71b09c92-6bce-42d0-aba1-7b985b8c3563";
+
+            firstPowerShellInstance.AddCommand("New-Item")
+                .AddParameter("Path", "Function:")
+                .AddParameter("Name", DummyFunctionName)
+                .AddParameter("Value", ScriptBlock.Create(string.Empty))
+                .InvokeAndClearCommands();
+
+            firstPowerShellInstance.AddCommand("Remove-Item")
+                .AddParameter("Path", $"Function:{DummyFunctionName}")
+                .InvokeAndClearCommands();
         }
 
         private static void LogPowerShellVersion(System.Management.Automation.PowerShell pwsh)